requirements.txt
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.10.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Session==0.5.0
Flask-WTF==1.2.1
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
passlib==1.7.4
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.0
Werkzeug==3.0.1
WTForms==3.1.2
zipp==3.17.0
